+++ src/objects/bullet.rs
@@ -177,8 +177,8 @@ impl GameEntity for Bullet{
         return self.collider.collides_with(other)
     }
 
-    fn get_collider(&self) -> Box<&dyn Collider> {
-        return Box::new(&self.collider)
+    fn get_collider(&self) -> &dyn Collider {
+        return &self.collider
     }
     
 }